pyproject.toml
[tool.poetry]
name = "hw-28"
version = "0.1.0"
description = ""
authors = ["p-and14 <p-and14@yandex.ru>"]
readme = "README.md"
packages = [{include = "hw_28"}]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.2"
psycopg2-binary = "^2.9.3"
Pillow = "^9.2.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.1"
pytest = "^7.1.3"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
